backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
azure-cosmos==4.6.0
azure-storage-blob==12.19.0 